Updated installers
Bug fixes

+++ janet-full-install.sh
@@ -192,12 +192,12 @@ systemctl start jarvis.service
 echo "-----------------------------------"
 
 echo "Borrando archivos temporales"
-rm -rf $DIRECTORY/Servidor/
-rm -rf $DIRECTORY/Cliente/
-rm -rf $DIRECTORY/Jarvis/
-rm $DIRECTORY/.gitignore
-rm $DIRECTORY/README.md
-rm $DIRECTORY/bibliotecas.json
+if [ -d "$DIRECTORY/Servidor" ]; then rm -Rf $DIRECTORY/Servidor; fi
+if [ -d "$DIRECTORY/Jarvis" ]; then rm -Rf $DIRECTORY/Jarvis; fi
+if [ -d "$DIRECTORY/Clientes" ]; then rm -Rf $DIRECTORY/Clientes; fi
+if [ -f "$DIRECTORY/.gitignore" ]; then rm $DIRECTORY/.gitignore; fi
+if [ -f "$DIRECTORY/README.md" ]; then rm $DIRECTORY/README.md; fi
+if [ -f "$DIRECTORY/LICENSE.md" ]; then rm $DIRECTORY/LICENSE.md; fi
 
 echo "Ok"
 echo "-----------------------------------"
